4-declarative-sidecar-pattern/generic_sidecar_engine/requirements.txt
redis==5.0.4
PyYAML==6.0.1